% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{checkCells}
\alias{checkCells}
\title{A function to check if cells are present in a Seurat object}
\usage{
checkCells(seurat_object, cells)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{features}{A vector of the feature names to be checked}
}
\description{
A function to check if cells are present in a Seurat object
}
