% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{checkFeatures}
\alias{checkFeatures}
\title{A function to check if features are present in a Seurat object}
\usage{
checkFeatures(seurat_object, features)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{features}{A vector of the feature names to be checked}
}
\description{
A function to check if features are present in a Seurat object
}
