% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.R
\name{convertHumanGeneList}
\alias{convertHumanGeneList}
\title{see: https://github.com/satijalab/seurat/issues/2493
and: https://www.r-bloggers.com/converting-mouse-to-human-gene-names-with-biomart-package/}
\usage{
convertHumanGeneList(human_gene_symbols)
}
\arguments{
\item{human_genes}{A vector of human gene symbols}
}
\description{
see: https://github.com/satijalab/seurat/issues/2493
and: https://www.r-bloggers.com/converting-mouse-to-human-gene-names-with-biomart-package/
}
