% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.R
\name{downsampleMatrix}
\alias{downsampleMatrix}
\title{Downsample a matrix of UMI counts,}
\usage{
downsampleMatrix(matrixUMI, downsample_method = "median", library_ids = c())
}
\arguments{
\item{matrixUMI}{sparse matrix of umi counts}

\item{downsample_method}{the statistic to normalise per-library}

\item{library_ids}{a vector of the library_ids}
}
\description{
Downsample a matrix of UMI counts,
}
