% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{expressionPlots}
\alias{expressionPlots}
\title{Draw a set of expression dotplots (2D version)}
\usage{
expressionPlots(
  seurat_object,
  features,
  rdims,
  x = "UMAP_1",
  y = "UMAP_2",
  ncol = 6,
  pch = 16,
  point_size = 2.5,
  max_quantile = 0.9
)
}
\arguments{
\item{features}{A vector of feature names}

\item{rdims}{A data frame or matrix containing the reduced dimensions}

\item{x}{The name of the rdims column to use for the x coordinates}

\item{y}{The name of the rdims column to use for the y coordinates}

\item{ncol}{The number of columns to use when facet wrapping}

\item{point_size}{The size of the points}

\item{max_quantile}{The expression quantile to cap the data at.}
}
\description{
Draw a set of expression dotplots (2D version)
}
