% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.R
\name{FastExpMeanChunked}
\alias{FastExpMeanChunked}
\title{Run FastExpMean by chunk to conserve memory}
\usage{
FastExpMeanChunked(x, rows_per_chunk = 2000)
}
\arguments{
\item{x}{a matrix}

\item{nrows}{number of rows per chunk}
}
\description{
Run FastExpMean by chunk to conserve memory
}
