% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Latex.R
\name{getFigureTex}
\alias{getFigureTex}
\title{Write latex for a figure}
\usage{
getFigureTex(image_file, caption, plot_dir_var = "plotDir", height = 0.9)
}
\arguments{
\item{image_file}{The image file to show in the figure.}

\item{caption}{The figure caption.}

\item{plot_dir_var}{The directory containing the image_file}

\item{height}{The height of the figure as a fraction of the page height.}
}
\description{
Write latex for a figure
}
