% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.R
\name{getHVG}
\alias{getHVG}
\title{Get highly variable genes using the trend var method from scran}
\usage{
getHVG(seurat_object, min_mean = 0, p_adjust_threshold = 0.05)
}
\arguments{
\item{seurat_object}{a seurat object}

\item{min_mean}{minimum mean counts threshold}

\item{p_adjust_threshold}{threshold for significance}
}
\description{
Get highly variable genes using the trend var method from scran
}
