% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{getSigPC}
\alias{getSigPC}
\title{Function to get the significant principle components following
a jackstraw analysis}
\usage{
getSigPC(
  seurat_object = NULL,
  ncomp = Inf,
  adjust_p = TRUE,
  p_adjust_method = "BH",
  p_threshold = 0.05,
  order_by_sig = FALSE
)
}
\arguments{
\item{seurat_object}{A seurat object on which JackStraw and JackStrawPlot have been run}

\item{ncomp}{If set, the max. number of significant components to return}

\item{adjust_p}{Should p values be adjusted}

\item{p_adjust_method}{Method for adjusting p values}

\item{p_threshold}{The significance threshold}

\item{order_by_sig}{If true, PCs are selected based on significance}
}
\description{
Function to get the significant principle components following
a jackstraw analysis
}
