% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Latex.R
\name{getSubFigureTex}
\alias{getSubFigureTex}
\title{Write latex for a subfigure}
\usage{
getSubFigureTex(image_file, caption, plot_dir_var = "plotDir", height = NULL)
}
\arguments{
\item{image_file}{The image file to show in the figure.}

\item{caption}{The figure caption.}

\item{plot_dir_var}{The directory containing the image_file}
}
\description{
Write latex for a subfigure
}
