% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Latex.R
\name{getSubsectionTex}
\alias{getSubsectionTex}
\title{Write latex for subsection title}
\usage{
getSubsectionTex(subsection_name)
}
\arguments{
\item{subsection_name}{The subsection title.}
}
\description{
Write latex for subsection title
}
