% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Latex.R
\name{getSubsubsectionTex}
\alias{getSubsubsectionTex}
\title{Write latex for subsubsection title}
\usage{
getSubsubsectionTex(subsubsection_name)
}
\arguments{
\item{subsubsection_name}{The subsubsection title.}
}
\description{
Write latex for subsubsection title
}
