% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinPlots.R
\name{getViolinData}
\alias{getViolinData}
\title{Get data for a set of violin plots}
\usage{
getViolinData(seurat_object, genes, metadata = NULL, clusters = NULL)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{genes}{The set of genes to retrieve data for (row names of the seurat objects' "data" slot).}

\item{metadata}{A vector of meta.data columns to retrieve}

\item{clusters}{A vector of clusters to retrieve data for, if NULL all are included.}
}
\description{
Get data for a set of violin plots
}
