% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{function to return colors matching ggplot2
see: https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of colours to return.}
}
\description{
function to return colors matching ggplot2
see: https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
