% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinPlots.R
\name{makeViolins}
\alias{makeViolins}
\title{Function to draw a block of horizontal violin lots}
\usage{
makeViolins(ggData, title = NULL, ncol = 8, group = NULL, colors = NULL,
  alpha = 1, clusters = NULL, xlab = NULL)
}
\arguments{
\item{ggData}{melted dataframe ready for plotting}

\item{title}{A title for this set of genes}

\item{ncol}{Number of columns, will be passed to facet_wrap}

\item{group}{A column on which to group plots by}

\item{colors}{A vector of (fill) colors, one per cluster.}

\item{alpha}{Alpha value for the fill color}

\item{clusters}{Optional list of clusters to plot}

\item{xlab}{The label for the x axis.}
}
\description{
The number of columns will be respected even if
there are fewer genes than columns
}
