% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{markerComplexHeatmap}
\alias{markerComplexHeatmap}
\title{A function to draw a heatmap of top cluster marker genes with
subgroup labels. The function uses the "scale.data" slot to make the heatmap.}
\usage{
markerComplexHeatmap(
  seurat_object,
  marker_table = NULL,
  n_markers = 20,
  cells_use = NULL,
  row_names_gp = 10,
  sub_group = NULL,
  disp_min = -2.5,
  disp_max = 2.5
)
}
\arguments{
\item{seurat_object}{A seurat objected with scaled data and cluster information}

\item{marker_table}{A dataframe containing the marker information. Must contain "cluster", "gene" and "avg_logFC" columns}

\item{n_markers}{The number of markers to plot}

\item{cells_use}{The names of the cells to use. If NULL all cells will be used}

\item{row_names_gp}{The font size for the gene names}

\item{sub_group}{If given, the name of a variable in the metadata of the seurat object}

\item{disp_min}{Disp floor, used to truncate the scaled data}

\item{disp_max}{Disp ceiling, used to truncate the scaled data}
}
\description{
A function to draw a heatmap of top cluster marker genes with
subgroup labels. The function uses the "scale.data" slot to make the heatmap.
}
