% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlluvialPlots.R
\name{plotAlluvial}
\alias{plotAlluvial}
\title{A function to generate alluvial plots for two cluster resolutions}
\usage{
plotAlluvial(
  path_clusters_left,
  path_clusters_right,
  name_left,
  name_right,
  prefix_left = NULL,
  prefix_right = NULL,
  min_cells = 0
)
}
\arguments{
\item{path_clusters_left}{A rds file with cluster assignments for each cell (as generated by seurat pipeline), to be plotted on the left side}

\item{path_clusters_right}{A rds file with cluster assignments for each cell (as generated by seurat pipeline), to be plotted on the right side}

\item{name_left}{Name for the left strata of the plot (e.g. resolution used); needs to be a character string}

\item{name_right}{Name for the right strata of the plot (e.g. resolution used); needs to be a character string}

\item{prefix_left}{Optional character string as prefix for left axis (if not used, the function will give a warning!)}

\item{prefix_right}{Optional character string as prefix for right axis (if not used, the function will give a warning!)}

\item{min_cells}{Minimum number of cells per connection ( > 0 by default).}
}
\value{
A ggplot object that can be plotted.
}
\description{
A function to generate alluvial plots for two cluster resolutions
}
\details{
Please note that names need to be character strings. Colors are chosen the same way as they are in pipeline_seurat.py and should correspond to the colors in the reports.
}
\examples{
plot_alluvial(path_clusters_left = "/path/to/seurat/all.seurat.dir/20_0.4_1_wilcox/cluster.dir/cluster_ids.rds", 
path_clusters_right =  "/path/to/seurat/all.seurat.dir/20_0.6_1_wilcox/cluster.dir/cluster_ids.rds", 
name_left = "clusters_20_0.4_1", name_right = "clusters_20_0.6_1")
}
