% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotFvE}
\alias{plotFvE}
\title{Make a plot of frequency versus expression level}
\usage{
plotFvE(
  data,
  p_col = "p.adj",
  label_col = "gene",
  m_col = "M",
  freq_col = "L",
  xlab = "change in expression (log2)",
  ylab = "change in frequency (log2)"
)
}
\arguments{
\item{data}{Matrix or dataframe}

\item{p_col}{Column containing p-values.}

\item{label_col}{Column containing labels for highlighted genes.}

\item{m_col}{Column containing M value (log2 exprs ratio)}

\item{freq_col}{Column containing L value (log2 freq ratio)}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}
}
\description{
Make a plot of frequency versus expression level
}
