% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinPlots.R
\name{plotHorizontalViolins}
\alias{plotHorizontalViolins}
\title{Make a plot (or return a grob) of a set of horizontal violin plots from a seurat object.}
\usage{
plotHorizontalViolins(seurat_object, genes, clusters = NULL, title = NULL,
  ncol = 8, xlab = "normalised expression level", group = NULL,
  colors = NULL, plot = TRUE)
}
\arguments{
\item{seurat_object}{A seurat object.}

\item{genes}{The list of genes to plot (should correspond to row names in the seurat object).}

\item{clusters}{The clusters to show - if NULL, plots are made for all clusters.}

\item{title}{A title for this set of genes.}

\item{ncol}{Number of columns, will be passed to facet_wrap.}

\item{xlab}{The label for the x axis.}

\item{group}{A column in the seurat "meta.data" slot on which to group plots by.}

\item{colors}{A vector of (fill) colors, one per cluster.}

\item{plot}{If set to FALSE the grob is returned instead.}
}
\description{
Make a plot (or return a grob) of a set of horizontal violin plots from a seurat object.
}
