% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotViolins}
\alias{plotViolins}
\title{Function to make sets of violin plots.}
\usage{
plotViolins(
  data,
  seurat_object,
  cluster_ids,
  type = "positive",
  group.by = NULL,
  ident.include = NULL,
  ncol = ncol,
  use.minfc = FALSE,
  minfc_col = "min_logFC",
  maxfc_col = "max_logFC",
  avgfc_col = "avg_logFC",
  m_col = "avg_logFC",
  p_col = "p.adj",
  id_col = "gene"
)
}
\arguments{
\item{data}{The data}

\item{seurat_object}{The seurat object}

\item{cluster_ids}{The cluster ids}

\item{type}{Either "positive" or "negative"}

\item{group.by}{A factor to group by}

\item{ident.include}{Identity to include}

\item{ncol}{Number of columns in the figure}

\item{use.minfc}{Use minimum foldchange}

\item{minfc_col}{Column containing the minimum foldchange}

\item{maxfc_col}{Column containing the maximum foldchage}

\item{avgfc_col}{Column containing the average foldchage}

\item{m_col}{Column containing M value (log2 exprs ratio)}

\item{p_col}{The column containing the p-value}

\item{id_col}{The column containing unique gene identifiers}
}
\description{
Function to make sets of violin plots.
}
