% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Make an volcano-style plot}
\usage{
plotVolcano(
  data,
  xlab = "xlab",
  ylab = "ylab",
  m_col = "M",
  a_col = "A",
  p_col = "p.adj",
  top_col = "top",
  label_col = "gene"
)
}
\arguments{
\item{data}{Matrix or dataframe}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{m_col}{Column containing M value (log2 ratio)}

\item{a_col}{Column containing A value (log2 mean expression)}

\item{p_col}{Column containing p-values.}

\item{top_col}{Column indicating top rows to highlight.}

\item{label_col}{Column containing labels for highlighted genes.}
}
\description{
Make an volcano-style plot
}
