% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.R
\name{Read10X}
\alias{Read10X}
\title{Slightly modified version of Read10x from Seurat.}
\usage{
Read10X(data.dir = NULL)
}
\arguments{
\item{data.dir}{The directory containing the 10x matrix, cell names and barcodes.}
}
\description{
Slightly modified version of Read10x from Seurat.
}
