% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlluvialPlots.R
\name{readClusterIDs}
\alias{readClusterIDs}
\title{A helper function to read in cluster_ids.rds files for alluvial plots}
\usage{
readClusterIDs(path)
}
\arguments{
\item{path}{A rds file with cluster assignments for each cell (as generated by seurat pipeline)}
}
\description{
A helper function to read in cluster_ids.rds files for alluvial plots
}
\details{
Helper function for alluvial diagrams to read in cluster assignments from cluster_ids.rds.
}
