% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinPlots.R
\name{removeGrobs}
\alias{removeGrobs}
\title{Remove spacer grobs from a facetted ggplot plot.}
\usage{
removeGrobs(gpGrob, ncol, nmissing)
}
\arguments{
\item{ncol}{The number of columns (as was passed to facet wrap)}

\item{nmissing}{The number of facet panels missing}

\item{gp}{The ggplot object containing the violin plots}
}
\description{
Remove spacer grobs from a facetted ggplot plot.
}
