% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{save_ggplots}
\alias{save_ggplots}
\title{Wrapper function for saving base R plots in both pdf and png format}
\usage{
save_ggplots(filepath = NULL, gp = NULL, width = 6, height = 8,
  dpi = 300, to_pdf = TRUE, to_png = TRUE)
}
\arguments{
\item{filepath}{The name of the file to write to (without .pdf or .png extension)}

\item{gp}{The ggplot object.}

\item{width}{The width of the plots.}

\item{height}{The height of the plots.}

\item{dpi}{The plot resolution.}

\item{to_pdf}{Save pdf version of plot (TRUE|FALSE).}

\item{to_png}{Save png version of plot (TRUE|FALSE).}
}
\description{
Wrapper function for saving base R plots in both pdf and png format
}
