% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{save_plots}
\alias{save_plots}
\title{Wrapper function for saving base R plots in both pdf and png format}
\usage{
save_plots(filepath = NULL, plot_fn = NULL, width = 6, height = 8,
  res = 300, to_pdf = TRUE, to_png = TRUE)
}
\arguments{
\item{filepath}{The name of the file to write to (without .pdf or .png extension)}

\item{plot_fn}{The plotting function to call.}

\item{width}{The width of the plots.}

\item{height}{The height of the plots.}

\item{res}{The plot resolution.}

\item{to_pdf}{Save pdf version of plot (TRUE|FALSE).}

\item{to_png}{Save png version of plot (TRUE|FALSE).}
}
\description{
Wrapper function for saving base R plots in both pdf and png format
}
