% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{scale_to_quantile}
\alias{scale_to_quantile}
\title{Rescale an expression matrix as a percentage of the given quantile}
\usage{
scale_to_quantile(x, q = 0.9)
}
\arguments{
\item{x}{the data}

\item{q}{the quantile to use as the maximum value}
}
\description{
Rescale an expression matrix as a percentage of the given quantile
}
