% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{sprintfResults}
\alias{sprintfResults}
\title{format numbers in results tables as character strings}
\usage{
sprintfResults(results_table, number_fmt = "\%0.3g")
}
\arguments{
\item{results_table}{A dataframe or matrix of results that contains some numeric columns.}

\item{number_fmt}{Sprintf format for numeric columns}
}
\description{
format numbers in results tables as character strings
}
