% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{topGenes}
\alias{topGenes}
\title{Function to add "top" indicator column to degenes}
\usage{
topGenes(data, m_col = "avg_log2FC", use_fc = TRUE, id_col = "gene", ngenes = 7)
}
\arguments{
\item{data}{The data}

\item{m_col}{The column containing the log2 ratio}

\item{id_col}{A column containing unique identifiers}

\item{ngenes}{The number of genes to demarcate}
}
\description{
Function to add "top" indicator column to degenes
}
