% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{violinPlotSection}
\alias{violinPlotSection}
\title{Make a latex section containing a set of violin plots}
\usage{
violinPlotSection(data, seurat_object, cluster_ids, type = "positive",
  group.by = opt$testfactor, ident.include = opt$identinclude,
  ncol = ncol, outdir = opt$outdir, analysis_title = "violin plots",
  fc_type = "fold change", use.minfc = FALSE)
}
\arguments{
\item{data}{The data}

\item{seurat_object}{The seurat object}

\item{cluster_ids}{The cluster ids}

\item{type}{Either "positive" or "negative"}

\item{group.by}{A factor to group by}

\item{ident.include}{Identity to include}

\item{ncol}{Number of columns in the figure}

\item{outdir}{The directory for the output}

\item{analysis_title}{Title for this section}

\item{fc_type}{The metric by which the violin plots are ordered}

\item{use.minfc}{Use minimum foldchange}
}
\description{
Make a latex section containing a set of violin plots
}
