% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Latex.R
\name{writeTex}
\alias{writeTex}
\title{Function for writing a tex snippet to a file}
\usage{
writeTex(file_path, tex)
}
\arguments{
\item{file_path}{The file to write the latex snippet to.}

\item{tex}{The tex snippet to write (a vector of lines).}
}
\description{
Function for writing a tex snippet to a file
}
