% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.R
\name{barcode2table}
\alias{barcode2table}
\title{Split barcode and aggregation identifier}
\usage{
barcode2table(codes)
}
\arguments{
\item{codes}{A chracter vector of cell barcodes.}
}
\value{
A data.frame of two columns:
}
\description{
Cell barcodes follow the format "barcode-aggegationId"
}
